requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.2.1
Werkzeug==3.1.3
qrcode==8.2
pillow==12.0.0
flask-talisman==1.1.0
